+++ al-khaser/Anti VM/Parallels.h
@@ -3,3 +3,4 @@
 
 
 VOID parallels_process();
+BOOL parallels_check_mac();