+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added checks for Hyper-V/Virtual-PC anti-VM (HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Virtual Machine\Guest\Parameters)
 - Added checks for multiple virtualization vendors using WMI (Select SerialNumber from Win32_BIOS).
 - Added checks for multiple virtualization vendors using WMI (Select Model/Manufacturer from Win32_ComputerSystem).
+- Added checks for MAC address for Xen, Parallels.
 
 #### 0.73
 - Bug fix: GetSystemFirmwareTable should take `resultBufferSize` as an argument for the second call.